a working program for creating a small telescope lens assembly
To use this program, run it like this:
$ perl telescope_rev_correctkerfed_mult2.pl  --svg --mult=3.56 > f.html

Creates an html file with embedded svg, viewable on modern browsers.  To
run this through Nick's laser I open this file in Internet Explorer(tm),
right click on the image and click "Save image as..." which will result
in IE creating an svg file.  (someday perhaps I need to produce an svg
file directly.)  The resulting file may be opened by Inkscape and
printed to the laser.  Alternatively one can print directly from IE (or
presumably any browser) to the laser driver/program.  The problem with
this is that IE produces an image that's 8.5x11 inches (smallest) which
messes up the laser's perimeter function.

The result is tight enough that it's possible to build this part of the
telescope without glue, though glue must be recommended for a perminant
scope of course.  This assembly fits snugly inside of paper tube with an
inner diameter of 1.5inches.