fixes unshifted starting point bug
In the Koike, by default, the beginning of every cut starts at 0,0. We
reflect that, but in the process of generating our shapes, if some kind
of translation has occurred to the initial part then the starting point
also shifts. Whereas this shift happens accurately within Koike::Part,
this shift has not been reflected properly in the generated GCode
because each GCode command specifies only where to go, assuming that the
present position has already been attained.  So, a little bit of extra
code is required in the GCode startup routine that adds a 'moveto'
command to the first valid sox,soy set of coordinates in the output.