adds an example file that generates a grid of hexagons
This file generates a grid of hexagons with optional nested hexagons.
As with all files created with Koike.pm, etc., this grid is designed to
generate working GCode for Bridgewire's plasma cutter, along with an SVG
analog for display. The SVG output is more useful than simply for
viewing, however, because it works well with Nick's software which uses
as its input for driving the grbl board.